remove v1 encoder and decoder in the engine
run-all-tests: true
