+++ daml-lf/encoder/src/test/scala/com/digitalasset/daml/lf/archive/testing/DamlLfEncoderTest.scala
@@ -9,7 +9,7 @@ import com.daml.daml_lf_dev.{DamlLf1, DamlLf2}
 import com.daml.lf.archive.{
   ArchivePayload,
   Dar,
-  DecodeCommon,
+  DecodeV2,
   UniversalArchiveDecoder,
   UniversalArchiveReader,
 }
@@ -36,7 +36,7 @@ class DamlLfEncoderTest
 
     "be readable" in {
 
-      val modules_1_8 = Set[DottedName](
+      val modules_2_1 = Set[DottedName](
         "UnitMod",
         "BoolMod",
         "Int64Mod",
@@ -56,23 +56,17 @@ class DamlLfEncoderTest
         "NumericMod",
         "AnyMod",
         "SynonymMod",
+        "GenMapMod",
+        "BigNumericMod",
+        "ExceptionMod",
+        "InterfaceMod",
+        "InterfaceMod0",
+        "InterfaceExtMod",
       )
-      val modules_1_11 = modules_1_8 + "GenMapMod"
-      val modules_1_13 = modules_1_11 + "BigNumericMod"
-      val modules_1_14 = modules_1_13 + "ExceptionMod"
-      val modules_1_15 = modules_1_14 + "InterfaceMod" + "InterfaceMod0"
-      val modules_1_dev = modules_1_15 + "InterfaceExtMod"
-      val modules_2_1 = modules_1_dev
       val modules_2_dev = modules_2_1
 
       val versions = Table(
         "versions" -> "modules",
-        "1.8" -> modules_1_8,
-        "1.11" -> modules_1_11,
-        "1.13" -> modules_1_13,
-        "1.14" -> modules_1_14,
-        "1.15" -> modules_1_15,
-        "1.dev" -> modules_1_dev,
         "2.1" -> modules_2_1,
         "2.dev" -> modules_2_dev,
       )
@@ -150,7 +144,7 @@ class DamlLfEncoderTest
     val builtinMod = ModuleName.assertFromString("BuiltinMod")
 
     "contains all builtins " in {
-      forEvery(Table("version", LanguageVersion.All.filter(LanguageVersion.v1_13 <= _): _*)) {
+      forEvery(Table("version", LanguageVersion.All.filter(LanguageVersion.v2_1 <= _): _*)) {
         // We do not check package older that 1.11 as they are used for stable packages only
         version =>
           val Right(dar) =
@@ -163,8 +157,8 @@ class DamlLfEncoderTest
             .values
             .collect { case Ast.DValue(_, Ast.EBuiltin(builtin), _) => builtin }
             .toSet
-          val builtinsInVersion = DecodeCommon.builtinFunctionInfos.collect {
-            case DecodeCommon.BuiltinFunctionInfo(_, builtin, minVersion, maxVersion, _)
+          val builtinsInVersion = DecodeV2.builtinFunctionInfos.collect {
+            case DecodeV2.BuiltinFunctionInfo(_, builtin, minVersion, maxVersion, _)
                 if minVersion <= version && maxVersion.forall(version < _) =>
               builtin
           }.toSet
