+++ language-support/java/codegen/BUILD.bazel
@@ -132,7 +132,7 @@ daml_compile(
 daml_compile(
     name = "test-daml-with-same-dependencies-but-different-target-version",
     srcs = ["src/test/daml/Foo2.daml"],
-    target = "1.13",
+    target = "2.dev",
 )
 
 daml_compile(
